requirements_basic.txt
# Basic FastAPI Requirements (No Rust dependencies at all)
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
sqlalchemy==1.4.53
pydantic==1.10.12

# Keep existing dependencies
openai==1.3.7
chromadb==0.4.18
pypdf2==3.0.1
streamlit==1.28.1 